refactor: clean up @mogami/solana by removing inline exports from main class
