+++ libs/solana/src/lib/interfaces/public-key-string.ts
@@ -1,0 +1,3 @@
+import { PublicKey } from '@solana/web3.js'
+
+export type PublicKeyString = PublicKey | string
