+++ libs/solana/src/lib/interfaces/solana-config.ts
@@ -1,0 +1,5 @@
+import { SolanaLogger } from './solana-logger'
+
+export interface SolanaConfig {
+  logger?: SolanaLogger
+}
