+++ libs/solana/src/lib/interfaces/token-balance.ts
@@ -1,0 +1,7 @@
+import BigNumber from 'bignumber.js'
+import { PublicKeyString } from './public-key-string'
+
+export interface TokenBalance {
+  account: PublicKeyString
+  balance: BigNumber
+}
